make no_std compatible
