+++ .travis.yml
@@ -10,7 +10,8 @@ script:
   - cargo build --verbose
   - cargo test --verbose
   - cargo test --verbose --features rustc-serialize
-  - cargo doc --no-deps
+  - if [[ `rustc -V` == *'nightly'* ]]; then cargo test --verbose --no-default-features --features nightly; fi
+  - cargo doc --no-deps 
 after_success:
   - travis-cargo --only nightly doc-upload
 env:
