+++ src/display.rs
@@ -1,4 +1,7 @@
+#[cfg(feature = "std")]
 use std::fmt::{self, Write};
+#[cfg(not(feature = "std"))]
+use core::fmt::{self, Write};
 
 use super::{TmFmt, Tm, Fmt};
