+++ src/lib.rs
@@ -29,12 +29,15 @@
 //! This crate uses the same syntax for format strings as the [strftime()]
 //! (http://man7.org/linux/man-pages/man3/strftime.3.html) function from the C
 //! standard library.
-
 #![doc(html_logo_url = "https://www.rust-lang.org/logos/rust-logo-128x128-blk-v2.png",
        html_favicon_url = "https://www.rust-lang.org/favicon.ico",
        html_root_url = "https://doc.rust-lang.org/time/")]
 #![allow(trivial_numeric_casts)]
+#![allow(unused_imports)]
 #![cfg_attr(test, deny(warnings))]
+#![cfg_attr(test, allow(deprecated))]
+#![cfg_attr(not(feature = "std"), no_std)]
+#![cfg_attr(feature = "nightly", feature(alloc))]
 
 #[cfg(target_os = "redox")] extern crate syscall;
 #[cfg(unix)] extern crate libc;
@@ -45,9 +48,28 @@
 #[cfg(test)] #[macro_use] extern crate log;
 #[cfg(all(windows, test))] extern crate advapi32;
 
+#[cfg(feature = "std")]
+mod std_impl;
+
+#[cfg(feature = "nightly")]
+#[macro_use]
+extern crate alloc;
+
+#[cfg(feature = "std")]
+pub use std_impl::*;
+
+#[cfg(not(feature = "std"))]
+use core::cmp::Ordering;
+#[cfg(not(feature = "std"))]
+use core::fmt;
+#[cfg(not(feature = "std"))]
+use core::ops::{Add, Sub};
+
+#[cfg(feature = "std")]
 use std::cmp::Ordering;
-use std::error::Error;
+#[cfg(feature = "std")]
 use std::fmt;
+#[cfg(feature = "std")]
 use std::ops::{Add, Sub};
 
 pub use duration::{Duration, OutOfRangeError};
@@ -63,6 +85,7 @@ pub use parse::strptime;
 mod display;
 mod duration;
 mod parse;
+
 mod sys;
 
 static NSEC_PER_SEC: i32 = 1_000_000_000;
@@ -144,142 +167,6 @@ impl Sub<Timespec> for Timespec {
     }
 }
 
-/**
- * Returns the current time as a `timespec` containing the seconds and
- * nanoseconds since 1970-01-01T00:00:00Z.
- */
-pub fn get_time() -> Timespec {
-    let (sec, nsec) = sys::get_time();
-    Timespec::new(sec, nsec)
-}
-
-
-/**
- * Returns the current value of a high-resolution performance counter
- * in nanoseconds since an unspecified epoch.
- */
-pub fn precise_time_ns() -> u64 {
-    sys::get_precise_ns()
-}
-
-
-/**
- * Returns the current value of a high-resolution performance counter
- * in seconds since an unspecified epoch.
- */
-pub fn precise_time_s() -> f64 {
-    return (precise_time_ns() as f64) / 1000000000.;
-}
-
-/// An opaque structure representing a moment in time.
-///
-/// The only operation that can be performed on a `PreciseTime` is the
-/// calculation of the `Duration` of time that lies between them.
-///
-/// # Examples
-///
-/// Repeatedly call a function for 1 second:
-///
-/// ```rust
-/// use time::{Duration, PreciseTime};
-/// # fn do_some_work() {}
-///
-/// let start = PreciseTime::now();
-///
-/// while start.to(PreciseTime::now()) < Duration::seconds(1) {
-///     do_some_work();
-/// }
-/// ```
-#[derive(Copy, Clone)]
-pub struct PreciseTime(u64);
-
-impl PreciseTime {
-    /// Returns a `PreciseTime` representing the current moment in time.
-    pub fn now() -> PreciseTime {
-        PreciseTime(precise_time_ns())
-    }
-
-    /// Returns a `Duration` representing the span of time from the value of
-    /// `self` to the value of `later`.
-    ///
-    /// # Notes
-    ///
-    /// If `later` represents a time before `self`, the result of this method
-    /// is unspecified.
-    ///
-    /// If `later` represents a time more than 293 years after `self`, the
-    /// result of this method is unspecified.
-    #[inline]
-    pub fn to(&self, later: PreciseTime) -> Duration {
-        // NB: even if later is less than self due to overflow, this will work
-        // since the subtraction will underflow properly as well.
-        //
-        // We could deal with the overflow when casting to an i64, but all that
-        // gets us is the ability to handle intervals of up to 584 years, which
-        // seems not very useful :)
-        Duration::nanoseconds((later.0 - self.0) as i64)
-    }
-}
-
-/// A structure representing a moment in time.
-///
-/// `SteadyTime`s are generated by a "steady" clock, that is, a clock which
-/// never experiences discontinuous jumps and for which time always flows at
-/// the same rate.
-///
-/// # Examples
-///
-/// Repeatedly call a function for 1 second:
-///
-/// ```rust
-/// # use time::{Duration, SteadyTime};
-/// # fn do_some_work() {}
-/// let start = SteadyTime::now();
-///
-/// while SteadyTime::now() - start < Duration::seconds(1) {
-///     do_some_work();
-/// }
-/// ```
-#[derive(Clone, Copy, PartialOrd, Ord, PartialEq, Eq, Debug)]
-pub struct SteadyTime(sys::SteadyTime);
-
-impl SteadyTime {
-    /// Returns a `SteadyTime` representing the current moment in time.
-    pub fn now() -> SteadyTime {
-        SteadyTime(sys::SteadyTime::now())
-    }
-}
-
-impl fmt::Display for SteadyTime {
-    fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
-        // TODO: needs a display customization
-        fmt::Debug::fmt(self, fmt)
-    }
-}
-
-impl Sub for SteadyTime {
-    type Output = Duration;
-
-    fn sub(self, other: SteadyTime) -> Duration {
-        self.0 - other.0
-    }
-}
-
-impl Sub<Duration> for SteadyTime {
-    type Output = SteadyTime;
-
-    fn sub(self, other: Duration) -> SteadyTime {
-        SteadyTime(self.0 - other)
-    }
-}
-
-impl Add<Duration> for SteadyTime {
-    type Output = SteadyTime;
-
-    fn add(self, other: Duration) -> SteadyTime {
-        SteadyTime(self.0 + other)
-    }
-}
 
 #[cfg(not(windows))]
 pub fn tzset() {
@@ -408,11 +295,6 @@ pub fn at_utc(clock: Timespec) -> Tm {
     tm
 }
 
-/// Returns the current time in UTC
-pub fn now_utc() -> Tm {
-    at_utc(get_time())
-}
-
 /// Returns the specified time in the local timezone
 pub fn at(clock: Timespec) -> Tm {
     let Timespec { sec, nsec } = clock;
@@ -422,10 +304,6 @@ pub fn at(clock: Timespec) -> Tm {
     tm
 }
 
-/// Returns the current time in the local timezone
-pub fn now() -> Tm {
-    at(get_time())
-}
 
 impl Tm {
     /// Convert time to the seconds from January 1, 1970
@@ -550,41 +428,6 @@ pub enum ParseError {
     UnexpectedCharacter(char, char),
 }
 
-impl fmt::Display for ParseError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
-            InvalidFormatSpecifier(ch) => {
-                write!(f, "{}: %{}", self.description(), ch)
-            }
-            UnexpectedCharacter(a, b) => {
-                write!(f, "expected: `{}`, found: `{}`", a, b)
-            }
-            _ => write!(f, "{}", self.description())
-        }
-    }
-}
-
-impl Error for ParseError {
-    fn description(&self) -> &str {
-        match *self {
-            InvalidSecond => "Invalid second.",
-            InvalidMinute => "Invalid minute.",
-            InvalidHour => "Invalid hour.",
-            InvalidDay => "Invalid day.",
-            InvalidMonth => "Invalid month.",
-            InvalidYear => "Invalid year.",
-            InvalidDayOfWeek => "Invalid day of the week.",
-            InvalidDayOfMonth => "Invalid day of the month.",
-            InvalidDayOfYear => "Invalid day of the year.",
-            InvalidZoneOffset => "Invalid zone offset.",
-            InvalidTime => "Invalid time.",
-            InvalidSecondsSinceEpoch => "Invalid seconds since epoch.",
-            MissingFormatConverter => "missing format converter after `%`",
-            InvalidFormatSpecifier(..) => "invalid format specifier",
-            UnexpectedCharacter(..) => "Unexpected character.",
-        }
-    }
-}
 
 /// A wrapper around a `Tm` and format string that implements Display.
 #[derive(Debug)]
@@ -641,26 +484,46 @@ fn validate_format<'a>(fmt: TmFmt<'a>) -> Result<TmFmt<'a>, ParseError> {
     Ok(fmt)
 }
 
-/// Formats the time according to the format string.
-pub fn strftime(format: &str, tm: &Tm) -> Result<String, ParseError> {
-    tm.strftime(format).map(|fmt| fmt.to_string())
-}
-
 #[cfg(test)]
 mod tests {
-    use super::{Timespec, get_time, precise_time_ns, precise_time_s,
-                at_utc, at, strptime, PreciseTime, SteadyTime, ParseError, Duration};
+    use super::{Timespec,
+                at_utc, at, strptime, ParseError, Duration};
     use super::ParseError::{InvalidTime, InvalidYear, MissingFormatConverter,
                             InvalidFormatSpecifier};
 
+
+    #[cfg(feature = "std")]
+    use super::{get_time};
+
+    #[cfg(feature = "std")]
+    use std::boxed::Box;
+
+    #[cfg(feature = "std")]
     use std::sync::{Once, ONCE_INIT, Mutex, MutexGuard, LockResult};
+    #[cfg(feature = "std")]
     use std::mem;
 
+    #[cfg(feature = "std")]
+    use std::string::ToString;
+
+    #[cfg(feature = "nightly")]
+    use alloc::string::ToString;
+
+    #[cfg(not(feature = "std"))]
+    use core::hash::{Hash, Hasher, SipHasher};
+    #[cfg(feature = "std")]
+    use std::hash::{Hash, Hasher, SipHasher};
+
+    #[cfg(feature = "std")]
     struct TzReset {
         _tzreset: ::sys::TzReset,
         _lock: LockResult<MutexGuard<'static, ()>>,
     }
 
+    #[cfg(not(feature = "std"))]
+    struct TzReset {}
+
+    #[cfg(feature = "std")]
     fn set_time_zone_la_or_london(london: bool) -> TzReset {
         // Lock manages current timezone because some tests require LA some
         // London
@@ -685,15 +548,23 @@ mod tests {
         }
     }
 
+    #[cfg(feature = "std")]
     fn set_time_zone() -> TzReset {
         set_time_zone_la_or_london(false)
     }
+    
+    #[cfg(not(feature = "std"))]
+    fn set_time_zone() -> TzReset {
+        TzReset {} 
+    }
 
+    #[cfg(feature = "std")]
     fn set_time_zone_london_dst() -> TzReset {
         set_time_zone_la_or_london(true)
     }
 
     #[test]
+    #[cfg(feature = "std")]
     fn test_get_time() {
         static SOME_RECENT_DATE: i64 = 1325376000i64; // 2012-01-01T00:00:00Z
         static SOME_FUTURE_DATE: i64 = 1577836800i64; // 2020-01-01T00:00:00Z
@@ -713,30 +584,6 @@ mod tests {
         if tv2.sec == tv1.sec {
             assert!(tv2.nsec >= tv1.nsec);
         }
-    }
-
-    #[test]
-    fn test_precise_time() {
-        let s0 = precise_time_s();
-        debug!("s0={} sec", s0);
-        assert!(s0 > 0.);
-
-        let ns0 = precise_time_ns();
-        let ns1 = precise_time_ns();
-        debug!("ns0={} ns", ns0);
-        debug!("ns1={} ns", ns1);
-        assert!(ns1 >= ns0);
-
-        let ns2 = precise_time_ns();
-        debug!("ns2={} ns", ns2);
-        assert!(ns2 >= ns1);
-    }
-
-    #[test]
-    fn test_precise_time_to() {
-        let t0 = PreciseTime(1000);
-        let t1 = PreciseTime(1023);
-        assert_eq!(Duration::nanoseconds(23), t0.to(t1));
     }
 
     #[test]
@@ -760,6 +607,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "std")]
     fn test_at() {
         let _reset = set_time_zone();
 
@@ -809,6 +657,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(any(feature = "std", feature = "nightly"))]
     fn test_strptime() {
         let _reset = set_time_zone();
 
@@ -834,7 +683,7 @@ mod tests {
                    Err(InvalidTime));
 
         match strptime("Fri Feb 13 15:31:30.01234 2009", format) {
-            Err(e) => panic!("{}", e),
+            Err(e) => panic!("{:?}", e),
             Ok(ref tm) => {
                 assert_eq!(tm.tm_sec, 30);
                 assert_eq!(tm.tm_min, 31);
@@ -1009,6 +858,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "std")]
     fn test_asctime() {
         let _reset = set_time_zone();
 
@@ -1023,6 +873,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "std")]
     fn test_ctime() {
         let _reset = set_time_zone();
 
@@ -1037,82 +888,95 @@ mod tests {
     }
 
     #[test]
-    fn test_strftime() {
+    #[cfg(any(feature = "nightly", feature = "std"))]
+    fn test_strftime_utc() {
+        let _reset = set_time_zone();
+
+        let time = Timespec::new(1234567890 - 8*3600, 54321);
+        let utc = at_utc(time);
+
+        assert_eq!(utc.strftime("").unwrap().to_string(), "".to_string());
+        assert_eq!(utc.strftime("%A").unwrap().to_string(), "Friday".to_string());
+        assert_eq!(utc.strftime("%a").unwrap().to_string(), "Fri".to_string());
+        assert_eq!(utc.strftime("%B").unwrap().to_string(), "February".to_string());
+        assert_eq!(utc.strftime("%b").unwrap().to_string(), "Feb".to_string());
+        assert_eq!(utc.strftime("%C").unwrap().to_string(), "20".to_string());
+        assert_eq!(utc.strftime("%c").unwrap().to_string(),
+                   "Fri Feb 13 15:31:30 2009".to_string());
+        assert_eq!(utc.strftime("%D").unwrap().to_string(), "02/13/09".to_string());
+        assert_eq!(utc.strftime("%d").unwrap().to_string(), "13".to_string());
+        assert_eq!(utc.strftime("%e").unwrap().to_string(), "13".to_string());
+        assert_eq!(utc.strftime("%F").unwrap().to_string(), "2009-02-13".to_string());
+        assert_eq!(utc.strftime("%f").unwrap().to_string(), "000054321".to_string());
+        assert_eq!(utc.strftime("%G").unwrap().to_string(), "2009".to_string());
+        assert_eq!(utc.strftime("%g").unwrap().to_string(), "09".to_string());
+        assert_eq!(utc.strftime("%H").unwrap().to_string(), "15".to_string());
+        assert_eq!(utc.strftime("%h").unwrap().to_string(), "Feb".to_string());
+        assert_eq!(utc.strftime("%I").unwrap().to_string(), "03".to_string());
+        assert_eq!(utc.strftime("%j").unwrap().to_string(), "044".to_string());
+        assert_eq!(utc.strftime("%k").unwrap().to_string(), "15".to_string());
+        assert_eq!(utc.strftime("%l").unwrap().to_string(), " 3".to_string());
+        assert_eq!(utc.strftime("%M").unwrap().to_string(), "31".to_string());
+        assert_eq!(utc.strftime("%m").unwrap().to_string(), "02".to_string());
+        assert_eq!(utc.strftime("%n").unwrap().to_string(), "\n".to_string());
+        assert_eq!(utc.strftime("%P").unwrap().to_string(), "pm".to_string());
+        assert_eq!(utc.strftime("%p").unwrap().to_string(), "PM".to_string());
+        assert_eq!(utc.strftime("%R").unwrap().to_string(), "15:31".to_string());
+        assert_eq!(utc.strftime("%r").unwrap().to_string(), "03:31:30 PM".to_string());
+        assert_eq!(utc.strftime("%S").unwrap().to_string(), "30".to_string());
+        assert_eq!(utc.strftime("%s").unwrap().to_string(), "1234539090".to_string());
+        assert_eq!(utc.strftime("%T").unwrap().to_string(), "15:31:30".to_string());
+        assert_eq!(utc.strftime("%t").unwrap().to_string(), "\t".to_string());
+        assert_eq!(utc.strftime("%U").unwrap().to_string(), "06".to_string());
+        assert_eq!(utc.strftime("%u").unwrap().to_string(), "5".to_string());
+        assert_eq!(utc.strftime("%V").unwrap().to_string(), "07".to_string());
+        assert_eq!(utc.strftime("%v").unwrap().to_string(), "13-Feb-2009".to_string());
+        assert_eq!(utc.strftime("%W").unwrap().to_string(), "06".to_string());
+        assert_eq!(utc.strftime("%w").unwrap().to_string(), "5".to_string());
+        // FIXME (#2350): support locale
+        assert_eq!(utc.strftime("%X").unwrap().to_string(), "15:31:30".to_string());
+        // FIXME (#2350): support locale
+        assert_eq!(utc.strftime("%x").unwrap().to_string(), "02/13/09".to_string());
+        assert_eq!(utc.strftime("%Y").unwrap().to_string(), "2009".to_string());
+        assert_eq!(utc.strftime("%y").unwrap().to_string(), "09".to_string());
+        // FIXME (#2350): support locale
+        assert_eq!(utc.strftime("%Z").unwrap().to_string(), "UTC".to_string());
+        assert_eq!(utc.strftime("%z").unwrap().to_string(), "-0000".to_string());
+        assert_eq!(utc.strftime("%+").unwrap().to_string(),
+                   "2009-02-13T15:31:30Z".to_string());
+        assert_eq!(utc.strftime("%%").unwrap().to_string(), "%".to_string());
+
+         let invalid_specifiers = ["%E", "%J", "%K", "%L", "%N", "%O", "%o", "%Q", "%q"];
+        for &sp in invalid_specifiers.iter() {
+            assert_eq!(utc.strftime(sp).unwrap_err(),
+                       InvalidFormatSpecifier(sp[1..].chars().next().unwrap()));
+        }
+        assert_eq!(utc.strftime("%").unwrap_err(), MissingFormatConverter);
+        assert_eq!(utc.strftime("%A %").unwrap_err(), MissingFormatConverter);
+        assert_eq!(utc.asctime().to_string(), "Fri Feb 13 15:31:30 2009".to_string());
+        assert_eq!(utc.rfc822z().to_string(), "Fri, 13 Feb 2009 15:31:30 -0000".to_string());
+        assert_eq!(utc.rfc3339().to_string(), "2009-02-13T15:31:30Z".to_string());
+
+    }
+
+    #[test]
+    #[cfg(feature = "std")]
+    fn test_strftime_local() {
         let _reset = set_time_zone();
 
         let time = Timespec::new(1234567890, 54321);
-        let utc = at_utc(time);
         let local = at(time);
 
-        assert_eq!(local.strftime("").unwrap().to_string(), "".to_string());
-        assert_eq!(local.strftime("%A").unwrap().to_string(), "Friday".to_string());
-        assert_eq!(local.strftime("%a").unwrap().to_string(), "Fri".to_string());
-        assert_eq!(local.strftime("%B").unwrap().to_string(), "February".to_string());
-        assert_eq!(local.strftime("%b").unwrap().to_string(), "Feb".to_string());
-        assert_eq!(local.strftime("%C").unwrap().to_string(), "20".to_string());
-        assert_eq!(local.strftime("%c").unwrap().to_string(),
-                   "Fri Feb 13 15:31:30 2009".to_string());
-        assert_eq!(local.strftime("%D").unwrap().to_string(), "02/13/09".to_string());
-        assert_eq!(local.strftime("%d").unwrap().to_string(), "13".to_string());
-        assert_eq!(local.strftime("%e").unwrap().to_string(), "13".to_string());
-        assert_eq!(local.strftime("%F").unwrap().to_string(), "2009-02-13".to_string());
-        assert_eq!(local.strftime("%f").unwrap().to_string(), "000054321".to_string());
-        assert_eq!(local.strftime("%G").unwrap().to_string(), "2009".to_string());
-        assert_eq!(local.strftime("%g").unwrap().to_string(), "09".to_string());
-        assert_eq!(local.strftime("%H").unwrap().to_string(), "15".to_string());
-        assert_eq!(local.strftime("%h").unwrap().to_string(), "Feb".to_string());
-        assert_eq!(local.strftime("%I").unwrap().to_string(), "03".to_string());
-        assert_eq!(local.strftime("%j").unwrap().to_string(), "044".to_string());
-        assert_eq!(local.strftime("%k").unwrap().to_string(), "15".to_string());
-        assert_eq!(local.strftime("%l").unwrap().to_string(), " 3".to_string());
-        assert_eq!(local.strftime("%M").unwrap().to_string(), "31".to_string());
-        assert_eq!(local.strftime("%m").unwrap().to_string(), "02".to_string());
-        assert_eq!(local.strftime("%n").unwrap().to_string(), "\n".to_string());
-        assert_eq!(local.strftime("%P").unwrap().to_string(), "pm".to_string());
-        assert_eq!(local.strftime("%p").unwrap().to_string(), "PM".to_string());
-        assert_eq!(local.strftime("%R").unwrap().to_string(), "15:31".to_string());
-        assert_eq!(local.strftime("%r").unwrap().to_string(), "03:31:30 PM".to_string());
-        assert_eq!(local.strftime("%S").unwrap().to_string(), "30".to_string());
-        assert_eq!(local.strftime("%s").unwrap().to_string(), "1234567890".to_string());
-        assert_eq!(local.strftime("%T").unwrap().to_string(), "15:31:30".to_string());
-        assert_eq!(local.strftime("%t").unwrap().to_string(), "\t".to_string());
-        assert_eq!(local.strftime("%U").unwrap().to_string(), "06".to_string());
-        assert_eq!(local.strftime("%u").unwrap().to_string(), "5".to_string());
-        assert_eq!(local.strftime("%V").unwrap().to_string(), "07".to_string());
-        assert_eq!(local.strftime("%v").unwrap().to_string(), "13-Feb-2009".to_string());
-        assert_eq!(local.strftime("%W").unwrap().to_string(), "06".to_string());
-        assert_eq!(local.strftime("%w").unwrap().to_string(), "5".to_string());
-        // FIXME (#2350): support locale
-        assert_eq!(local.strftime("%X").unwrap().to_string(), "15:31:30".to_string());
-        // FIXME (#2350): support locale
-        assert_eq!(local.strftime("%x").unwrap().to_string(), "02/13/09".to_string());
-        assert_eq!(local.strftime("%Y").unwrap().to_string(), "2009".to_string());
-        assert_eq!(local.strftime("%y").unwrap().to_string(), "09".to_string());
-        // FIXME (#2350): support locale
+        assert_eq!(local.ctime().to_string(), "Fri Feb 13 15:31:30 2009".to_string());
+        assert_eq!(local.asctime().to_string(), "Fri Feb 13 15:31:30 2009".to_string());
+        assert_eq!(local.ctime().to_string(), "Fri Feb 13 15:31:30 2009".to_string());
+        assert_eq!(local.rfc822().to_string(), "Fri, 13 Feb 2009 15:31:30 ".to_string());
+        assert_eq!(local.rfc822z().to_string(), "Fri, 13 Feb 2009 15:31:30 -0800".to_string());
+        assert_eq!(local.rfc3339().to_string(), "2009-02-13T15:31:30-08:00".to_string());
         assert_eq!(local.strftime("%Z").unwrap().to_string(), "".to_string());
         assert_eq!(local.strftime("%z").unwrap().to_string(), "-0800".to_string());
         assert_eq!(local.strftime("%+").unwrap().to_string(),
                    "2009-02-13T15:31:30-08:00".to_string());
-        assert_eq!(local.strftime("%%").unwrap().to_string(), "%".to_string());
-
-         let invalid_specifiers = ["%E", "%J", "%K", "%L", "%N", "%O", "%o", "%Q", "%q"];
-        for &sp in invalid_specifiers.iter() {
-            assert_eq!(local.strftime(sp).unwrap_err(),
-                       InvalidFormatSpecifier(sp[1..].chars().next().unwrap()));
-        }
-        assert_eq!(local.strftime("%").unwrap_err(), MissingFormatConverter);
-        assert_eq!(local.strftime("%A %").unwrap_err(), MissingFormatConverter);
-
-        assert_eq!(local.asctime().to_string(), "Fri Feb 13 15:31:30 2009".to_string());
-        assert_eq!(local.ctime().to_string(), "Fri Feb 13 15:31:30 2009".to_string());
-        assert_eq!(local.rfc822z().to_string(), "Fri, 13 Feb 2009 15:31:30 -0800".to_string());
-        assert_eq!(local.rfc3339().to_string(), "2009-02-13T15:31:30-08:00".to_string());
-
-        assert_eq!(utc.asctime().to_string(), "Fri Feb 13 23:31:30 2009".to_string());
-        assert_eq!(utc.ctime().to_string(), "Fri Feb 13 15:31:30 2009".to_string());
-        assert_eq!(utc.rfc822().to_string(), "Fri, 13 Feb 2009 23:31:30 GMT".to_string());
-        assert_eq!(utc.rfc822z().to_string(), "Fri, 13 Feb 2009 23:31:30 -0000".to_string());
-        assert_eq!(utc.rfc3339().to_string(), "2009-02-13T23:31:30Z".to_string());
     }
 
     #[test]
@@ -1150,27 +1014,26 @@ mod tests {
     #[test]
     #[allow(deprecated)]
     fn test_timespec_hash() {
-        use std::hash::{Hash, Hasher};
 
         let c = &Timespec::new(3, 2);
         let d = &Timespec::new(2, 1);
         let e = &Timespec::new(2, 1);
 
-        let mut hasher = ::std::hash::SipHasher::new();
+        let mut hasher = SipHasher::new();
 
         let d_hash:u64 = {
           d.hash(&mut hasher);
           hasher.finish()
         };
 
-        hasher = ::std::hash::SipHasher::new();
+        hasher = SipHasher::new();
 
         let e_hash:u64 = {
           e.hash(&mut hasher);
           hasher.finish()
         };
 
-        hasher = ::std::hash::SipHasher::new();
+        hasher = SipHasher::new();
 
         let c_hash:u64 = {
           c.hash(&mut hasher);
@@ -1228,18 +1091,10 @@ mod tests {
 
     #[test]
     fn test_time_sub() {
-        let a = ::now();
+        let a = ::empty_tm();
         let b = at(a.to_timespec() + Duration::seconds(5));
         let c = b - a;
         assert_eq!(c.num_nanoseconds(), Some(super::NSEC_PER_SEC as i64 * 5));
-    }
-
-    #[test]
-    fn test_steadytime_sub() {
-        let a = SteadyTime::now();
-        let b = a + Duration::seconds(1);
-        assert_eq!(b - a, Duration::seconds(1));
-        assert_eq!(a - b, Duration::seconds(-1));
     }
 
     #[test]
@@ -1250,6 +1105,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "std")]
     fn test_dst() {
         let _reset = set_time_zone_london_dst();
         let utc_in_feb = strptime("2015-02-01Z", "%F%z").unwrap();
